Made security functions return NDN_SUCCESS (defined in ndn-error-code.h) rather than 0, changed comments to reflect this, fixed bug in hmac signing and verification functions where input and key values were passed in wrong order to hmac_sha256 function (which is why it still passed unit tests)

+++ app-support/bootstrapping/secure-sign-on-files/secure-sign-on/tlv/sign-on-basic-tlv-helpers.c
@@ -12,8 +12,8 @@
 
 #include "sign-on-basic-tlv-impl-consts.h"
 
-enum ParseTlvValueResultCode parseTlvValue(const uint8_t *tlvBlock, uint16_t tlvBlockLength, int tlvTypeToFindValueOf, 
-                        uint16_t *tlvValueLength, uint16_t *tlvValueOffset) {
+enum ParseTlvValueResultCode parseTlvValue(const uint8_t *tlvBlock, uint32_t tlvBlockLength, int tlvTypeToFindValueOf, 
+                        uint32_t *tlvValueLength, uint32_t *tlvValueOffset) {
 
   int i = 0;
