+++ app-support/bootstrapping/secure-sign-on-files/secure-sign-on/variants/basic/security/detail/detail-aes/aes-nrf-crypto-impl.h
@@ -40,9 +40,9 @@
 
 #define NRF_CRYPTO_EXAMPLE_AES_MAX_TEXT_SIZE 120
 
-int sign_on_basic_nrf_crypto_decrypt_aes_cbc_pkcs5pad(uint8_t *key, uint16_t key_len, 
-    const uint8_t *encrypted_payload, uint16_t encrypted_payload_len,
-    uint8_t *decrypted_payload, uint16_t *decrypted_payload_len);
+int sign_on_basic_nrf_crypto_decrypt_aes_cbc_pkcs5pad(uint8_t *key, uint32_t key_len, 
+    const uint8_t *encrypted_payload, uint32_t encrypted_payload_len,
+    uint8_t *decrypted_payload, uint32_t *decrypted_payload_len);
 
 //*****************************************************//
