+++ app-support/bootstrapping/secure-sign-on-files/secure-sign-on/variants/basic/security/detail/sec-impl.h
@@ -20,31 +20,31 @@
 // since many boards will share the same security libraries, but for now I will do it this way...
 #define nRF52840
 
-int sign_on_basic_gen_sha256_hash(const uint8_t *payload, uint16_t payload_len, uint8_t *output);
+int sign_on_basic_gen_sha256_hash(const uint8_t *payload, uint32_t payload_len, uint8_t *output);
 
-int sign_on_basic_decrypt_aes_cbc_pkcs5pad(uint8_t *key, uint16_t key_len, 
-                                           const uint8_t *encrypted_payload, uint16_t encrypted_payload_len,
-                                           uint8_t *decrypted_payload, uint16_t *decrypted_payload_len);
+int sign_on_basic_decrypt_aes_cbc_pkcs5pad(uint8_t *key, uint32_t key_len, 
+                                           const uint8_t *encrypted_payload, uint32_t encrypted_payload_len,
+                                           uint8_t *decrypted_payload, uint32_t *decrypted_payload_len);
 
-int sign_on_basic_vrfy_hmac_sha256_sig(const uint8_t *payload, uint16_t payload_len,
-                                                  const uint8_t *sig, uint16_t sig_len,
-                                                  const uint8_t *key, uint16_t key_len);
+int sign_on_basic_vrfy_hmac_sha256_sig(const uint8_t *payload, uint32_t payload_len,
+                                                  const uint8_t *sig, uint32_t sig_len,
+                                                  const uint8_t *key, uint32_t key_len);
 
 int sign_on_basic_gen_sha256_ecdsa_sig(const uint8_t *pri_key_raw,
-                                                const uint8_t *payload, uint16_t payload_len,
-                                                uint8_t *output_buf, uint16_t output_buf_len, 
-                                                uint16_t *output_len);
+                                                const uint8_t *payload, uint32_t payload_len,
+                                                uint8_t *output_buf, uint32_t output_buf_len, 
+                                                uint32_t *output_len);
 
-int sign_on_basic_gen_ecdh_shared_secret(const uint8_t *pub_key_raw, uint16_t pub_key_raw_len,
-                                                    const uint8_t *pri_key_raw, uint16_t pri_key_raw_len,
+int sign_on_basic_gen_ecdh_shared_secret(const uint8_t *pub_key_raw, uint32_t pub_key_raw_len,
+                                                    const uint8_t *pri_key_raw, uint32_t pri_key_raw_len,
                                                     uECC_Curve curve,
-                                                    uint8_t *output_buf, uint16_t output_buf_len, 
-                                                    uint16_t *output_len);
+                                                    uint8_t *output_buf, uint32_t output_buf_len, 
+                                                    uint32_t *output_len);
 
-int sign_on_basic_gen_ec_keypair(uint8_t *pub_key_buf, uint16_t pub_key_buf_len, 
-                                            uint16_t *pub_key_output_len,
-                                            uint8_t *pri_key_buf, uint16_t pri_key_buf_len, 
-                                            uint16_t *pri_key_output_len,
+int sign_on_basic_gen_ec_keypair(uint8_t *pub_key_buf, uint32_t pub_key_buf_len, 
+                                            uint32_t *pub_key_output_len,
+                                            uint8_t *pri_key_buf, uint32_t pri_key_buf_len, 
+                                            uint32_t *pri_key_output_len,
                                             uECC_Curve curve);
 
 #endif // SIGN_ON_BLE_BASIC_SEC_IMPL_H
